test(administration): split the order state history modal spec
The spec had grown to 725 lines, past the 500-line threshold at which
`adr/2026-05-06-split-large-administration-test-files.md` asks for a split.

Group the cases by behaviour and move the shared fixtures and the wrapper
factory into a non-spec helper so Jest does not run them as tests. The wrapper
resolves the component itself, following the existing split specs, which drops
the per-file `beforeAll`. No assertion changed.
